Add location custom hook
